+++ mongoengine/fields.py
@@ -1082,8 +1082,8 @@ class GridFSProxy(object):
 
     def __repr__(self):
         return '<%s: %s>' % (self.__class__.__name__, self.grid_id)
-        
-    def __unicode__(self):
+
+    def __str__(self):
         name = getattr(self.get(), 'filename', self.grid_id) if self.get() else '(no file)'
         return '<%s: %s>' % (self.__class__.__name__, name)