+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 Changes in 0.8.4
 ================
+- Added str representation of GridFSProxy (#424)
 - Update transform to handle docs erroneously passed to unset (#416)
 - Fixed indexing - turn off _cls (#414)
 - Fixed dereference threading issue in ComplexField.__get__ (#412)